+++ allennlp/common/file_utils.py
@@ -48,15 +48,15 @@ if os.path.exists(DEPRECATED_CACHE_DIRECTORY):
     )
 
 
-def url_to_filename(url: str, etag: str = None) -> str:
+def _resource_to_filename(resource: str, etag: str = None) -> str:
     """
-    Convert `url` into a hashed filename in a repeatable way.
-    If `etag` is specified, append its hash to the url's, delimited
+    Convert a `resource` into a hashed filename in a repeatable way.
+    If `etag` is specified, append its hash to the resources's, delimited
     by a period.
     """
-    url_bytes = url.encode("utf-8")
-    url_hash = sha256(url_bytes)
-    filename = url_hash.hexdigest()
+    resource_bytes = resource.encode("utf-8")
+    resource_hash = sha256(resource_bytes)
+    filename = resource_hash.hexdigest()
 
     if etag:
         etag_bytes = etag.encode("utf-8")
@@ -121,6 +121,8 @@ def cached_path(
     if cache_dir is None:
         cache_dir = CACHE_DIRECTORY
 
+    os.makedirs(cache_dir, exist_ok=True)
+
     if isinstance(url_or_filename, PathLike):
         url_or_filename = str(url_or_filename)
 
@@ -166,11 +168,13 @@ def cached_path(
         file_path = url_or_filename
 
         if extract_archive and (is_zipfile(file_path) or tarfile.is_tarfile(file_path)):
-            # This is the path the file should be extracted to.
-            # For example model.tar.gz -> model-tar-gz-extracted
-            extraction_dir, extraction_name = os.path.split(file_path)
-            extraction_name = extraction_name.replace(".", "-") + "-extracted"
-            extraction_path = os.path.join(extraction_dir, extraction_name)
+            # We'll use a unique directory within the cache to root to extract the archive to.
+            # The name of the directoy is a hash of the resource file path and it's modification
+            # time. That way, if the file changes, we'll know when to extract it again.
+            extraction_name = (
+                _resource_to_filename(file_path, str(os.path.getmtime(file_path))) + "-extracted"
+            )
+            extraction_path = os.path.join(cache_dir, extraction_name)
 
     elif parsed.scheme == "":
         # File, but it doesn't exist.
@@ -187,7 +191,7 @@ def cached_path(
             return extraction_path
 
         # Extract it.
-        with FileLock(file_path + ".lock"):
+        with FileLock(extraction_path + ".lock"):
             shutil.rmtree(extraction_path, ignore_errors=True)
 
             # We extract first to a temporary directory in case something goes wrong
@@ -326,7 +330,7 @@ def _http_get(url: str, temp_file: IO) -> None:
 
 
 def _find_latest_cached(url: str, cache_dir: Union[str, Path]) -> Optional[str]:
-    filename = url_to_filename(url)
+    filename = _resource_to_filename(url)
     cache_path = os.path.join(cache_dir, filename)
     candidates: List[Tuple[str, float]] = []
     for path in glob.glob(cache_path + "*"):
@@ -424,8 +428,6 @@ def get_from_cache(url: str, cache_dir: Union[str, Path] = None) -> str:
     if cache_dir is None:
         cache_dir = CACHE_DIRECTORY
 
-    os.makedirs(cache_dir, exist_ok=True)
-
     # Get eTag to add to filename, if it exists.
     try:
         if url.startswith("s3://"):
@@ -463,7 +465,7 @@ def get_from_cache(url: str, cache_dir: Union[str, Path] = None) -> str:
         # If this is the case, try to proceed without eTag check.
         etag = None
 
-    filename = url_to_filename(url, etag)
+    filename = _resource_to_filename(url, etag)
 
     # Get cache path to put the file.
     cache_path = os.path.join(cache_dir, filename)